BayesianPINN Discretizer for PDESystems

Using the Bayesian PINN solvers, we can solve general nonlinear PDEs, ODEs and also simultaneously perform parameter estimation on them.

Note: The BPINN PDE solver also works for ODEs defined using ModelingToolkit, ModelingToolkit.jl PDESystem documentation. Despite this, the ODE specific BPINN solver BNNODE refer exists and uses NeuralPDE.ahmc_bayesian_pinn_ode at a lower level.

BayesianPINN Discretizer for PDESystems and lower level Bayesian PINN Solver calls for PDEs and ODEs.

NeuralPDE.BayesianPINN — Type

BayesianPINN(chain,
                  strategy;
                  init_params = nothing,
                  phi = nothing,
                  param_estim = false,
                  additional_loss = nothing,
                  adaptive_loss = nothing,
                  logger = nothing,
                  log_options = LogOptions(),
                  iteration = nothing,
                  dataset = nothing,
                  kwargs...) where {iip}

A discretize algorithm for the ModelingToolkit PDESystem interface, which transforms a PDESystem into a likelihood function used for HMC based Posterior Sampling Algorithms AdvancedHMC.jl which is later optimized upon to give the Solution Distribution of the PDE, using the Physics-Informed Neural Networks (PINN) methodology.

Positional Arguments

chain: a vector of Flux.jl or Lux.jl chains with a d-dimensional input and a 1-dimensional output corresponding to each of the dependent variables. Note that this specification respects the order of the dependent variables as specified in the PDESystem.

strategy: determines which training strategy will be used. See the Training Strategy documentation for more details.

Keyword Arguments

Dataset: A vector of matrix, each matrix for ith dependant variable and first col in matrix is for dependant variables, remaining coloumns for independant variables. Needed for inverse problem solving.

init_params: the initial parameters of the neural networks. This should match the specification of the chosen chain library. For example, if a Flux.chain is used, then init_params should match Flux.destructure(chain)[1] in shape. If init_params is not given, then the neural network default parameters are used. Note that for Lux, the default will convert to Float64.

phi: a trial solution, specified as phi(x,p) where x is the coordinates vector for the dependent variable and p are the weights of the phi function (generally the weights of the neural network defining phi). By default, this is generated from the chain. This should only be used to more directly impose functional information in the training problem, for example imposing the boundary condition by the test function formulation.

adaptive_loss: (STILL WIP), the choice for the adaptive loss function. See the adaptive loss page for more details. Defaults to no adaptivity.

additional_loss: a function additional_loss(phi, θ, p_) where phi are the neural network trial solutions, θ are the weights of the neural network(s), and p_ are the hyperparameters . If param_estim = true, then θ additionally contains the parameters of the differential equation appended to the end of the vector.

param_estim: whether the parameters of the differential equation should be included in the values sent to the additional_loss function. Defaults to false.

logger: ?? needs docs

log_options: ?? why is this separate from the logger?

iteration: used to control the iteration counter???

kwargs: Extra keyword arguments.

NeuralPDE.ahmc_bayesian_pinn_ode — Function

priors: pdf for W,b + pdf for ODE params

NeuralPDE.ahmc_bayesian_pinn_pde — Function

priors: pdf for W,b + pdf for PDE params

symbolic_discretize for BayesianPINN and lower level interface.

SciMLBase.symbolic_discretize — Method

prob = symbolic_discretize(pde_system::PDESystem, discretization::AbstractPINN)

symbolic_discretize is the lower level interface to discretize for inspecting internals. It transforms a symbolic description of a ModelingToolkit-defined PDESystem into a PINNRepresentation which holds the pieces required to build an OptimizationProblem for Optimization.jl or a Likelihood Function used for HMC based Posterior Sampling Algorithms AdvancedHMC.jl which is later optimized upon to give Solution or the Solution Distribution of the PDE.

For more information, see discretize and PINNRepresentation.

NeuralPDE.BPINNstats — Type

Contains ahmcbayesianpinnode() function output: 1> a MCMCChains.jl chain object for sampled parameters 2> The set of all sampled parameters 3> statistics like: > nsteps > acceptancerate > logdensity > hamiltonianenergy > hamiltonianenergyerror > numericalerror > stepsize > nomstep_size

NeuralPDE.BPINNsolution — Type

BPINN Solution contains the original solution from AdvancedHMC.jl sampling(BPINNstats contains fields related to that)

ensemblesol is the Probabilistic Estimate(MonteCarloMeasurements.jl Particles type) of Ensemble solution from All Neural Network's(made using all sampled parameters) output's. estimatednnparams - Probabilistic Estimate of NN params from sampled weights,biases estimateddeparams - Probabilistic Estimate of DE params from sampled unknown DE paramters