BayesianPINN Discretizer for PDESystems

Using the Bayesian PINN solvers, we can solve general nonlinear PDEs, ODEs and also simultaneously perform parameter estimation on them.

Note: The BPINN PDE solver also works for ODEs defined using ModelingToolkit, ModelingToolkit.jl PDESystem documentation. Despite this, the ODE specific BPINN solver BNNODE refer exists and uses NeuralPDE.ahmc_bayesian_pinn_ode at a lower level.

BayesianPINN Discretizer for PDESystems and lower level Bayesian PINN Solver calls for PDEs and ODEs.

NeuralPDE.BayesianPINN — Type

BayesianPINN(chain,
              strategy;
              init_params = nothing,
              phi = nothing,
              param_estim = false,
              additional_loss = nothing,
              adaptive_loss = nothing,
              logger = nothing,
              log_options = LogOptions(),
              iteration = nothing,
              dataset = nothing,
              kwargs...)

A discretize algorithm for the ModelingToolkit PDESystem interface, which transforms a PDESystem into a likelihood function used for HMC based Posterior Sampling Algorithms AdvancedHMC.jl which is later optimized upon to give the Solution Distribution of the PDE, using the Physics-Informed Neural Networks (PINN) methodology.

Positional Arguments

chain: a vector of Lux.jl chains with a d-dimensional input and a 1-dimensional output corresponding to each of the dependent variables. Note that this specification respects the order of the dependent variables as specified in the PDESystem.

strategy: determines which training strategy will be used. See the Training Strategy documentation for more details.

Keyword Arguments

Dataset: A vector of matrix, each matrix for ith dependant variable and first col in matrix is for dependant variables, remaining columns for independent variables. Needed for inverse problem solving.

init_params: the initial parameters of the neural networks. If init_params is not given, then the neural network default parameters are used. Note that for Lux, the default will convert to Float64.

phi: a trial solution, specified as phi(x,p) where x is the coordinates vector for the dependent variable and p are the weights of the phi function (generally the weights of the neural network defining phi). By default, this is generated from the chain. This should only be used to more directly impose functional information in the training problem, for example imposing the boundary condition by the test function formulation.

adaptive_loss: (STILL WIP), the choice for the adaptive loss function. See the adaptive loss page for more details. Defaults to no adaptivity.

additional_loss: a function additional_loss(phi, θ, p_) where phi are the neural network trial solutions, θ are the weights of the neural network(s), and p_ are the hyperparameters . If param_estim = true, then θ additionally contains the parameters of the differential equation appended to the end of the vector.

param_estim: whether the parameters of the differential equation should be included in the values sent to the additional_loss function. Defaults to false.

logger: ?? needs docs

log_options: ?? why is this separate from the logger?

iteration: used to control the iteration counter???

kwargs: Extra keyword arguments.

NeuralPDE.ahmc_bayesian_pinn_ode — Function

ahmc_bayesian_pinn_ode(prob, chain; strategy = GridTraining,
                    dataset = [nothing],init_params = nothing, 
                    draw_samples = 1000, physdt = 1 / 20.0f0,l2std = [0.05],
                    phystd = [0.05], priorsNNw = (0.0, 2.0),
                    param = [], nchains = 1, autodiff = false, Kernel = HMC,
                    Adaptorkwargs = (Adaptor = StanHMCAdaptor,
                                     Metric = DiagEuclideanMetric, 
                                     targetacceptancerate = 0.8),
                    Integratorkwargs = (Integrator = Leapfrog,),
                    MCMCkwargs = (n_leapfrog = 30,),
                    progress = false, verbose = false)

Warn

Note that ahmcbayesianpinnode() only supports ODEs which are written in the out-of-place form, i.e. du = f(u,p,t), and not f(du,u,p,t). If not declared out-of-place, then the ahmcbayesianpinnode() will exit with an error.

Example

linear = (u, p, t) -> -u / p[1] + exp(t / p[2]) * cos(t)
tspan = (0.0, 10.0)
u0 = 0.0
p = [5.0, -5.0]
prob = ODEProblem(linear, u0, tspan, p)

### CREATE DATASET (Necessity for accurate Parameter estimation)
sol = solve(prob, Tsit5(); saveat = 0.05)
u = sol.u[1:100]
time = sol.t[1:100]

### dataset and BPINN create
x̂ = collect(Float64, Array(u) + 0.05 * randn(size(u)))
dataset = [x̂, time]

chain1 = Lux.Chain(Lux.Dense(1, 5, tanh), Lux.Dense(5, 5, tanh), Lux.Dense(5, 1)

### simply solving ode here hence better to not pass dataset(uses ode params specified in prob)
fh_mcmc_chain1, fhsamples1, fhstats1 = ahmc_bayesian_pinn_ode(prob, chain1,
                                                            dataset = dataset,
                                                            draw_samples = 1500,
                                                            l2std = [0.05],
                                                            phystd = [0.05],
                                                            priorsNNw = (0.0,3.0))

### solving ode + estimating parameters hence dataset needed to optimize parameters upon + Pior Distributions for ODE params
fh_mcmc_chain2, fhsamples2, fhstats2 = ahmc_bayesian_pinn_ode(prob, chain1,
                                                            dataset = dataset,
                                                            draw_samples = 1500,
                                                            l2std = [0.05],
                                                            phystd = [0.05],
                                                            priorsNNw = (0.0,3.0),
                                                            param = [Normal(6.5,0.5), Normal(-3,0.5)])

NOTES

Dataset is required for accurate Parameter estimation + solving equations Incase you are only solving the Equations for solution, do not provide dataset

Positional Arguments

prob: DEProblem(out of place and the function signature should be f(u,p,t).

chain: Lux Neural Netork which would be made the Bayesian PINN.

Keyword Arguments

strategy: The training strategy used to choose the points for the evaluations. By default GridTraining is used with given physdt discretization.

init_params: initial parameter values for BPINN (ideally for multiple chains different initializations preferred)

nchains: number of chains you want to sample

draw_samples: number of samples to be drawn in the MCMC algorithms (warmup samples are ~2/3 of draw samples)

l2std: standard deviation of BPINN prediction against L2 losses/Dataset

phystd: standard deviation of BPINN prediction against Chosen Underlying ODE System

priorsNNw: Tuple of (mean, std) for BPINN Network parameters. Weights and Biases of BPINN are Normal Distributions by default.

param: Vector of chosen ODE parameters Distributions in case of Inverse problems.

autodiff: Boolean Value for choice of Derivative Backend(default is numerical)

physdt: Timestep for approximating ODE in it's Time domain. (1/20.0 by default)

Kernel: Choice of MCMC Sampling Algorithm (AdvancedHMC.jl implementations HMC/NUTS/HMCDA)

Integratorkwargs: Integrator, jitter_rate, tempering_rate. Refer: https://turinglang.org/AdvancedHMC.jl/stable/

Adaptorkwargs: Adaptor, Metric, targetacceptancerate. Refer: https://turinglang.org/AdvancedHMC.jl/stable/ Note: Target percentage(in decimal) of iterations in which the proposals are accepted (0.8 by default)

MCMCargs: A NamedTuple containing all the chosen MCMC kernel's(HMC/NUTS/HMCDA) Arguments, as follows :

n_leapfrog: number of leapfrog steps for HMC

δ: target acceptance probability for NUTS and HMCDA

λ: target trajectory length for HMCDA

max_depth: Maximum doubling tree depth (NUTS)

Δ_max: Maximum divergence during doubling tree (NUTS)

Refer: https://turinglang.org/AdvancedHMC.jl/stable/

progress: controls whether to show the progress meter or not.

verbose: controls the verbosity. (Sample call args in AHMC)

Warnings

AdvancedHMC.jl is still developing convenience structs so might need changes on new releases.

NeuralPDE.ahmc_bayesian_pinn_pde — Function

ahmc_bayesian_pinn_pde(pde_system, discretization;
        draw_samples = 1000,
        bcstd = [0.01], l2std = [0.05],
        phystd = [0.05], priorsNNw = (0.0, 2.0),
        param = [], nchains = 1, Kernel = HMC(0.1, 30),
        Adaptorkwargs = (Adaptor = StanHMCAdaptor,
            Metric = DiagEuclideanMetric, targetacceptancerate = 0.8),
        Integratorkwargs = (Integrator = Leapfrog,), saveats = [1 / 10.0],
        numensemble = floor(Int, draw_samples / 3), progress = false, verbose = false)

NOTES

Dataset is required for accurate Parameter estimation + solving equations.

Returned solution is a BPINNsolution consisting of Ensemble solution, estimated PDE and NN parameters for chosen saveats grid spacing and last n = numensemble samples in Chain. the complete set of samples in the MCMC chain is returned as fullsolution, refer BPINNsolution for more details.

Positional Arguments

pde_system: ModelingToolkit defined PDE equation or system of equations.

discretization: BayesianPINN discretization for the given pde_system, Neural Network and training strategy.

Keyword Arguments

draw_samples: number of samples to be drawn in the MCMC algorithms (warmup samples are ~2/3 of draw samples)

bcstd: Vector of standard deviations of BPINN prediction against Initial/Boundary Condition equations.

l2std: Vector of standard deviations of BPINN prediction against L2 losses/Dataset for each dependant variable of interest.

phystd: Vector of standard deviations of BPINN prediction against Chosen Underlying PDE equations.

priorsNNw: Tuple of (mean, std) for BPINN Network parameters. Weights and Biases of BPINN are Normal Distributions by default.

param: Vector of chosen PDE's parameter's Distributions in case of Inverse problems.

nchains: number of chains you want to sample.

Kernel: Choice of MCMC Sampling Algorithm object HMC/NUTS/HMCDA (AdvancedHMC.jl implementations).

Adaptorkwargs: Adaptor, Metric, targetacceptancerate. Refer: https://turinglang.org/AdvancedHMC.jl/stable/ Note: Target percentage(in decimal) of iterations in which the proposals are accepted (0.8 by default).

Integratorkwargs: Integrator, jitter_rate, tempering_rate. Refer: https://turinglang.org/AdvancedHMC.jl/stable/

saveats: Grid spacing for each independent variable for evaluation of ensemble solution, estimated parameters.

numensemble: Number of last samples to take for creation of ensemble solution, estimated parameters.

progress: controls whether to show the progress meter or not.

verbose: controls the verbosity. (Sample call args in AHMC).

Warnings

AdvancedHMC.jl is still developing convenience structs so might need changes on new releases.

symbolic_discretize for BayesianPINN and lower level interface.

SciMLBase.symbolic_discretize — Method

prob = symbolic_discretize(pde_system::PDESystem, discretization::AbstractPINN)

symbolic_discretize is the lower level interface to discretize for inspecting internals. It transforms a symbolic description of a ModelingToolkit-defined PDESystem into a PINNRepresentation which holds the pieces required to build an OptimizationProblem for Optimization.jl or a Likelihood Function used for HMC based Posterior Sampling Algorithms AdvancedHMC.jl which is later optimized upon to give Solution or the Solution Distribution of the PDE.

For more information, see discretize and PINNRepresentation.

NeuralPDE.BPINNstats — Type

Contains ahmcbayesianpinnode() function output: 1> a MCMCChains.jl chain object for sampled parameters 2> The set of all sampled parameters 3> statistics like: > nsteps > acceptancerate > logdensity > hamiltonianenergy > hamiltonianenergyerror > numericalerror > stepsize > nomstep_size

NeuralPDE.BPINNsolution — Type

BPINN Solution contains the original solution from AdvancedHMC.jl sampling(BPINNstats contains fields related to that)

ensemblesol is the Probabilistic Estimate(MonteCarloMeasurements.jl Particles type) of Ensemble solution from All Neural Network's(made using all sampled parameters) output's. estimatednnparams - Probabilistic Estimate of NN params from sampled weights,biases estimateddeparams - Probabilistic Estimate of DE params from sampled unknown DE parameters